requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
sqlalchemy>=2.0.27
psycopg[binary]>=3.1.0
pydantic>=2.6.0
python-dotenv>=1.0.0
streamlit>=1.31.0
requests>=2.31.0
openai>=1.12.0
anthropic>=0.18.0
pandas>=2.2.0
plotly>=5.18.0
pytest>=8.0.0
